[ADD] Base line models

+++ src/plants_sm/featurization/proteins/encodings/nlf.py
@@ -58,4 +58,4 @@ class NLFEncoder(FeaturesGenerator):
         np.ndarray: The encoded sequence.
 
         """
-        return np.array([np.array(NLF[aa]) for aa in sequence])
+        return np.array([np.array(NLF[aa]) for aa in sequence], dtype=np.float32)
